Use SubstrateCoinEntry trait to better support Substrate chains.
