+++ rust/tw_coin_registry/Cargo.toml
@@ -26,6 +26,7 @@ tw_misc = { path = "../tw_misc" }
 tw_native_evmos = { path = "../chains/tw_native_evmos" }
 tw_native_injective = { path = "../chains/tw_native_injective" }
 tw_pactus = { path = "../chains/tw_pactus" }
+tw_substrate = { path = "../tw_substrate" }
 tw_polkadot = { path = "../chains/tw_polkadot" }
 tw_ronin = { path = "../chains/tw_ronin" }
 tw_solana = { path = "../chains/tw_solana" }
