+++ rust/tw_substrate/Cargo.toml
@@ -11,3 +11,4 @@ tw_hash = { path = "../tw_hash" }
 tw_keypair = { path = "../tw_keypair" }
 tw_memory = { path = "../tw_memory" }
 tw_ss58_address = { path = "../tw_ss58_address" }
+tw_proto = { path = "../tw_proto" }
